refactor: Move to TS and update sentence-splitter/to-string
