+++ package.json
@@ -2,18 +2,18 @@
   "name": "textlint-rule-en-max-word-count",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/azu/textlint-rule-en-max-word-count.git"
+    "url": "git+https://github.com/textlint-rule/textlint-rule-en-max-word-count.git"
   },
   "author": "azu",
   "email": "azuciao@gmail.com",
-  "homepage": "https://github.com/azu/textlint-rule-en-max-word-count",
+  "homepage": "https://github.com/textlint-rule/textlint-rule-en-max-word-count",
   "license": "MIT",
   "files": [
     "src/",
     "lib/"
   ],
   "bugs": {
-    "url": "https://github.com/azu/textlint-rule-en-max-word-count/issues"
+    "url": "https://github.com/textlint-rule/textlint-rule-en-max-word-count/issues"
   },
   "version": "2.0.0",
   "description": "textlint rule that specify the maximum word count of a sentence.",
@@ -22,13 +22,16 @@
     "test": "test"
   },
   "dependencies": {
-    "sentence-splitter": "^3.2.2",
-    "textlint-tester": "^12.2.0",
-    "textlint-util-to-string": "^3.1.1",
-    "unist-util-map": "^1.0.1"
+    "sentence-splitter": "^4.1.0",
+    "textlint-tester": "^13.3.0",
+    "textlint-util-to-string": "^3.3.2"
   },
   "devDependencies": {
-    "textlint-scripts": "^12.2.1"
+    "@types/node": "^18.13.0",
+    "@textlint/types": "^13.3.0",
+    "textlint-scripts": "^13.3.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5"
   },
   "peerDependencies": {
     "textlint": ">= 12.2.0"
